Make breaking channel hierarchy changes
* (Re)move methods deprecated in preparation for threads
* Move most base channel method implementations to default methods
* Remove most base channel classes
