+++ core/src/main/java/discord4j/core/object/entity/channel/NewsChannel.java
@@ -31,7 +31,7 @@ import java.util.Objects;
 import java.util.function.Consumer;
 
 /** A Discord news channel. */
-public final class NewsChannel extends BaseTopLevelGuildMessageChannel {
+public final class NewsChannel extends BaseTopLevelGuildChannel implements TopLevelGuildMessageChannel {
 
     /**
      * Constructs an {@code NewsChannel} with an associated {@link GatewayDiscordClient} and Discord data.
